Replaced make-project.sh and the old sync script with a new project creating, management and deployment system.
make-project.sh and sync are replaced by the afk script, a thin wrapper
around libsync.sh, which does initial minimal project setup and command
dispatching.

The meat is the new libsync.sh library. It runs different functions in
response to subcommands in an extensible manner and contains the rest of
the project setup, deployment, and management code.

Making new projects should just be a matter of copying the afk script
into the directory to contain the project, running it while telling it
where to find AFK, and that's it! As all but the barest functionality
has been moved to a library, the afk script itself should remain stable
for the forseeable future and any changes that end up being made to the
system can now be done in a sane manner to libsync.sh instead, making
everything much more maintainable.

The project template is now a bunch of plain files rather than a bunch
of heredocs, which is also a huge improvement over the former situation.